requirements.txt
Flask
Pillow
numpy
scipy
